requirements.txt
asgiref==3.11.0
Django==5.2.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.4
tzdata==2025.2
